requirements.txt
anthropic>=0.30.0
python-dotenv>=1.0.1
pypdf>=4.2.0
